Standardbudget Repository Pattern implementiert
- Für Standardbudget das Repository Pattern implementiert.
- Kleinere Code Verbesserungen implementiert Funktionsnamen wurden verbessert.
