+++ lib/models/default_budget/default_budget_interface.dart
@@ -1,0 +1,8 @@
+import 'default_budget_model.dart';
+
+abstract class DefaultBudgetInterface {
+  void create(DefaultBudget newDefaultBudget);
+  void update(DefaultBudget updatedDefaultBudget);
+  void delete(String budgetCategorie);
+  Future<DefaultBudget> load(String defaultBudgetCategorie);
+}
