+++ lib/models/default_budget/default_budget_repository.dart
@@ -2,20 +2,18 @@ import 'package:hive/hive.dart';
 
 import '/utils/consts/hive_consts.dart';
 
-@HiveType(typeId: defaultBudgetTypeId)
-class DefaultBudget extends HiveObject {
-  late int boxIndex;
-  @HiveField(0)
-  late String categorie;
-  @HiveField(1)
-  late double defaultBudget;
+import 'default_budget_interface.dart';
+import 'default_budget_model.dart';
 
-  void createDefaultBudget(DefaultBudget newDefaultBudget) async {
+class DefaultBudgetRepository extends DefaultBudgetInterface {
+  @override
+  void create(DefaultBudget newDefaultBudget) async {
     var defaultBudgetBox = await Hive.openBox(defaultBudgetsBox);
     defaultBudgetBox.add(newDefaultBudget);
   }
 
-  void updateDefaultBudget(DefaultBudget updatedDefaultBudget) async {
+  @override
+  void update(DefaultBudget updatedDefaultBudget) async {
     var defaultBudgetBox = await Hive.openBox(defaultBudgetsBox);
     for (int i = 0; i < defaultBudgetBox.length; i++) {
       DefaultBudget defaultBudget = await defaultBudgetBox.getAt(i);
@@ -25,7 +23,8 @@ class DefaultBudget extends HiveObject {
     }
   }
 
-  static void deleteDefaultBudget(String budgetCategorie) async {
+  @override
+  void delete(String budgetCategorie) async {
     var defaultBudgetBox = await Hive.openBox(defaultBudgetsBox);
     for (int i = 0; i < defaultBudgetBox.length; i++) {
       DefaultBudget defaultBudget = await defaultBudgetBox.getAt(i);
@@ -36,7 +35,8 @@ class DefaultBudget extends HiveObject {
     }
   }
 
-  static Future<DefaultBudget> loadDefaultBudget(String defaultBudgetCategorie) async {
+  @override
+  Future<DefaultBudget> load(String defaultBudgetCategorie) async {
     var defaultBudgetBox = await Hive.openBox(defaultBudgetsBox);
     DefaultBudget defaultBudget = DefaultBudget()
       ..categorie = ''
@@ -48,23 +48,5 @@ class DefaultBudget extends HiveObject {
       }
     }
     return defaultBudget;
-  }
-}
-
-class DefaultBudgetAdapter extends TypeAdapter<DefaultBudget> {
-  @override
-  final typeId = defaultBudgetTypeId;
-
-  @override
-  DefaultBudget read(BinaryReader reader) {
-    return DefaultBudget()
-      ..categorie = reader.read()
-      ..defaultBudget = reader.read();
-  }
-
-  @override
-  void write(BinaryWriter writer, DefaultBudget obj) {
-    writer.write(obj.categorie);
-    writer.write(obj.defaultBudget);
   }
 }
