+++ lib/screens/edit_subbudget_screen.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:haushaltsbuch/models/subbudget/subbudget_repository.dart';
 
-import '/models/default_budget.dart';
 import '/models/subbudget/subbudget_model.dart';
+import '/models/subbudget/subbudget_repository.dart';
+import '/models/default_budget/default_budget_model.dart';
+import '/models/default_budget/default_budget_repository.dart';
 
 import '/utils/consts/route_consts.dart';
 
@@ -35,7 +36,8 @@ class _EditSubbudgetScreenState extends State<EditSubbudgetScreen> {
 
   Future<List<Subbudget>> _loadOneSubbudgetCategorie() async {
     SubbudgetRepository subbudgetRepository = SubbudgetRepository();
-    _defaultBudget = await DefaultBudget.loadDefaultBudget(widget.subbudget.subcategorieName);
+    DefaultBudgetRepository defaultBudgetRepository = DefaultBudgetRepository();
+    _defaultBudget = await defaultBudgetRepository.load(widget.subbudget.subcategorieName);
     _subbudgetList = await subbudgetRepository.loadOneSubbudget(widget.subbudget.subcategorieName);
     return _subbudgetList;
   }
